+++ gsonpath-compiler/src/test/java/gsonpath/generator/adapter/auto/FieldValidationTypeTest.kt
@@ -7,6 +7,7 @@ class FieldValidationTypeTest : BaseGeneratorTest() {
     @Test
     fun testValidateExplicitNonNull() {
         assertGeneratedContent(BaseGeneratorTest.TestCriteria("adapter/auto/field_policy/validate_explicit_non_null")
+                .addAbsoluteSource("adapter/auto/TestGsonTypeFactory.java")
                 .addRelativeSource("TestValidateExplicitNonNull.java")
                 .addRelativeGenerated("TestValidateExplicitNonNull_GsonTypeAdapter.java"))
     }
@@ -14,6 +15,7 @@ class FieldValidationTypeTest : BaseGeneratorTest() {
     @Test
     fun testValidateAllExceptNullable() {
         assertGeneratedContent(BaseGeneratorTest.TestCriteria("adapter/auto/field_policy/validate_all_except_nullable")
+                .addAbsoluteSource("adapter/auto/TestGsonTypeFactory.java")
                 .addRelativeSource("TestValidateAllExceptNullable.java")
                 .addRelativeGenerated("TestValidateAllExceptNullable_GsonTypeAdapter.java"))
     }
@@ -21,6 +23,7 @@ class FieldValidationTypeTest : BaseGeneratorTest() {
     @Test
     fun testNoValidation() {
         assertGeneratedContent(BaseGeneratorTest.TestCriteria("adapter/auto/field_policy/no_validation")
+                .addAbsoluteSource("adapter/auto/TestGsonTypeFactory.java")
                 .addRelativeSource("TestNoValidation.java")
                 .addRelativeGenerated("TestNoValidation_GsonTypeAdapter.java"))
     }