+++ gsonpath-compiler/src/test/java/gsonpath/generator/adapter/auto/SerializeNullsTest.kt
@@ -8,6 +8,7 @@ class SerializeNullsTest : BaseGeneratorTest() {
     @Test
     fun testSerializeNulls() {
         assertGeneratedContent(BaseGeneratorTest.TestCriteria("adapter/auto/serialize_nulls")
+                .addAbsoluteSource("adapter/auto/TestGsonTypeFactory.java")
                 .addRelativeSource("TestSerializeNulls.java")
                 .addRelativeGenerated("TestSerializeNulls_GsonTypeAdapter.java"))
     }