+++ gsonpath/src/main/java/gsonpath/GsonPath.java
@@ -1,6 +1,6 @@
 package gsonpath;
 
-import gsonpath.internal.GsonPathTypeAdapterFactory;
+import com.google.gson.TypeAdapterFactory;
 
 /**
  * The primary class to use when using the GsonPath library.
@@ -8,8 +8,10 @@ import gsonpath.internal.GsonPathTypeAdapterFactory;
  * It supplies factories which expose auto generated class created using the {@link gsonpath.AutoGsonAdapter} annotation.
  */
 public class GsonPath {
+    private static final String FACTORY_IMPLEMENTATION_SUFFIX = "Impl";
+
     /**
-     * Creates an instance of the {@link gsonpath.internal.GsonPathTypeAdapterFactory} class.
+     * Creates an instance of an {@link TypeAdapterFactory} implementation class that implements the input interface.
      * <p>
      * This factory is used to map the auto generated {@link com.google.gson.TypeAdapter} classes created using the
      * {@link gsonpath.AutoGsonAdapter} annotation.
@@ -17,9 +19,16 @@ public class GsonPath {
      * Only a single use of reflection is used within the constructor, so it isn't critical to hold onto this reference
      * for later usage.
      *
-     * @return a new instance of the {@link gsonpath.internal.GsonPathTypeAdapterFactory} class
+     * @param clazz the type adatper class to use to find the concrete implementation. Ensure that the interface is
+     *              annotated with {@link gsonpath.AutoGsonAdapterFactory}
+     * @return a new instance of the {@link TypeAdapterFactory} class
      */
-    public static GsonPathTypeAdapterFactory createTypeAdapterFactory() {
-        return new GsonPathTypeAdapterFactory();
+    public static TypeAdapterFactory createTypeAdapterFactory(Class<? extends TypeAdapterFactory> clazz) {
+        String factoryClassName = clazz.getCanonicalName() + FACTORY_IMPLEMENTATION_SUFFIX;
+        try {
+            return (TypeAdapterFactory) Class.forName(factoryClassName).newInstance();
+        } catch (Exception e) {
+            throw new RuntimeException("Unable to find generated TypeAdapterFactory '" + factoryClassName + "'");
+        }
     }
 }