+++ CHANGELOG.md
@@ -1,6 +1,31 @@
 Change Log
 ==========
 
+Version 2.0.0-beta1 *(2017-02-13)*
+----------------------------
+
+* New: Added multi-module support. Previously the library would not allow gsonpath to be used with a library.
+     * This is a breaking API change, to use the library a new `@AutoGsonAdapterFactory` must be used.
+
+     Previously, the factory was added to Gson as follows:
+
+     `builder.registerTypeAdapterFactory(GsonPath.createTypeAdapterFactory())`
+
+     The following step is now required beforehand:
+
+     ```java
+     @AutoGsonAdapterFactory
+     public interface TestGsonTypeFactory extends TypeAdapterFactory {
+     }
+     ```
+
+     Also, notice that the `createTypeAdapterFactory` has now requires a class as an argument.
+
+     `builder.registerTypeAdapterFactory(GsonPath.createTypeAdapterFactory(TestGsonTypeFactory.class))`
+
+     The advantage of this change is that multiple `@AutoGsonAdapterFactory` annotations can be used.
+     *Note: Only one `@AutoGsonAdapterFactory` annotation may be used per module/project*
+
 Version 1.8.0 *(2017-01-08)*
 ----------------------------